@@ -5,14 +5,14 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="nbxplorer.client" Version="2.0.0.6" />
-    <PackageReference Include="Nethereum.Geth" Version="3.4.0" />
-    <PackageReference Include="RavenDB.Embedded" Version="4.2.2" />
-    <PackageReference Include="Serilog" Version="2.8.0" />
+    <PackageReference Include="NBXplorer.Client" Version="3.0.1" />
+    <PackageReference Include="Nethereum.Geth" Version="3.5.0" />
+    <PackageReference Include="RavenDB.Embedded" Version="4.2.8" />
+    <PackageReference Include="Serilog" Version="2.9.0" />
   </ItemGroup>
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <LangVersion>7.3</LangVersion>
+    <TargetFramework>netstandard2.1</TargetFramework>
+    <LangVersion>8</LangVersion>
   </PropertyGroup>
 </Project>
