@@ -31,7 +31,7 @@ namespace Bizanc.io.Matching.Api
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc();
+            services.AddControllers();
 
             services.AddSwaggerGen(c =>
             {
@@ -60,13 +60,8 @@ namespace Bizanc.io.Matching.Api
             services.AddLogging(l => l.AddSerilog(dispose: true, logger: Log.Logger));
         }
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, Microsoft.AspNetCore.Hosting.IWebHostEnvironment env)
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
             app.UseSwagger();
             app.UseSwaggerUI(c =>
             {
@@ -81,7 +76,14 @@ namespace Bizanc.io.Matching.Api
                 builder.AllowAnyHeader();
                 builder.AllowAnyMethod();
             });
-            app.UseMvc();
+            //app.UseMvc();
+
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
